+++ frontend/src/Componentes/Pagina/InfoTonto/ListaParticipantes.jsx
@@ -1,22 +1,8 @@
-import { useEffect, useState } from "react";
-import { Api } from "../../Api/Api";
 import ListaTitulos from "./ListaTitulos";
+import { useNombres } from "../../Hooks/useNombres";
 
 const ListaParticipantes = () => {
-  // https://thecowboys.duckdns.org/api/cowboys
-  // https://reqres.in/api/users?page=2
-
-  const [users, setUsers] = useState([]);
-
-  useEffect(() => {
-    getUsers();
-  }, []);
-
-  const getUsers = async () => {
-    const resp = await Api.get("https://reqres.in/api/users?page=2");
-    setUsers(resp.data.data);
-  };
-
+  const { users } = useNombres();
   return (
     <>
       {users.map((user) => (
